pyproject.toml
[tool.poetry]
name = "cinema"
version = "0.1.0"
description = ""
authors = ["adamjanas <a.janas@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
isort = "^5.4.1"
flake8 = "^3.8.3"
bandit = "^1.6.2"
autoflake = "^1.3.1"
django-cleanup = "^5.0.0"
django-crispy-forms = "^1.9.2"
django-phone-field = "^1.8.1"
xhtml2pdf = "^0.2.4"
autopep8 = "^1.5.4"

[tool.isort]
line_length = 120
known_future_library = "future"
known_first_party = "app"
known_django = "django"
known_third_party = "django-cleanup, django-crispy-forms, django-phone-field, xhtml2pdf"
multi_line_output = 3
include_trailing_comma = "True"
indent = '    '
skip = "manage.py,migrations,venv,node_modules,virtualenv"
sections = "FUTURE, STDLIB, DJANGO, THIRDPARTY, FIRSTPARTY, LOCALFOLDER"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
